converter, dtos and petservices requests implemented on controller layer

+++ src/main/java/com/petbackend/api/builder/PetBuilder.java
@@ -8,7 +8,7 @@ import com.petbackend.api.model.PetType;
 @Component
 public class PetBuilder {
 	
-	public static Pet build(Integer id, String name, PetType type, Integer age) {
+	public static Pet build(Long id, String name, PetType type, Integer age) {
 		return Pet.builder().id(id).name(name).type(type).age(age).build();
 	}
